refactor(Prize, Prizes, LadderGame): 피드백 반영 - 메소드 인자로 int 를 받아서 column을 새로 생성하는 것에서 Column을 인자로 받도록 변경

+++ src/main/java/laddergame/domain/game/LadderGame.java
@@ -20,7 +20,7 @@ public class LadderGame {
 
         for (Player player : players.getPlayers()) {
             Column afterMoveColumn = ladder.progressAllStep(player.getColumn());
-            resultMap.put(player.getName(), prizes.findByColumn(afterMoveColumn.toInt()));
+            resultMap.put(player.getName(), prizes.findByColumn(afterMoveColumn));
         }
 
         return new GameResult(resultMap);
